pyproject.toml
[tool.poetry]
name = "auto-video-platform"
version = "0.1.0"
description = "Skeleton проекта для системы автоматизации медиаконтента"
authors = ["Your Team <dev@example.com>"]
readme = "README.md"
packages = [
  { include = "apps" },
  { include = "api" },
  { include = "workflows" },
  { include = "domain" },
  { include = "infrastructure" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.6.0"
pydantic-settings = "^2.2.1"
sqlalchemy = { version = "^2.0.29", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.1"
structlog = "^24.1.0"
httpx = "^0.27.0"
redis = "^5.0.4"
boto3 = "^1.34.0"
minio = "^7.2.6"
tenacity = "^8.2.3"
sentry-sdk = "^1.45.0"
prometheus-client = "^0.20.0"
temporalio = "^1.4.0"
authlib = "^1.3.0"
python-dotenv = "^1.0.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
bcrypt = "<5.0.0"
email-validator = "^2.2.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
ruff = "^0.4.1"
isort = "^5.13.2"
mypy = "^1.9.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
extend-select = ["B", "C4", "Q", "I", "UP"]

[tool.pytest.ini_options]
addopts = "--strict-markers --disable-warnings"
asyncio_mode = "auto"
